Done fixing smart contracts

+++ contracts/interfaces/IERC4907.sol
@@ -6,14 +6,14 @@ interface IERC4907 {
     // Logged when the user of a token assigns a new user or updates expires
     /// @notice Emitted when the `user` of an NFT or the `expires` of the `user` is changed
     /// The zero address for user indicates that there is no user address
-    event UpdateUser(uint256 indexed tokenId, address indexed user, uint256 expires);
+    event UpdateUser(uint256 indexed tokenId, address indexed user, uint96 expires);
 
     /// @notice set the user and expires of a NFT
     /// @dev The zero address indicates there is no user
     /// Throws if `tokenId` is not valid NFT
     /// @param user  The new user of the NFT
     /// @param expires  UNIX timestamp, The new user could use the NFT before expires
-    function setUser(uint256 tokenId, address user, uint256 expires) external;
+    function setUser(uint256 tokenId, address user, uint96 expires) external;
 
     /// @notice Get the user address of an NFT
     /// @dev The zero address indicates that there is no user or the user is expired
@@ -25,5 +25,5 @@ interface IERC4907 {
     /// @dev The zero value indicates that there is no user
     /// @param tokenId The NFT to get the user expires for
     /// @return The user expires for this NFT
-    function userExpires(uint256 tokenId) external view returns(uint256);
+    function userExpires(uint256 tokenId) external view returns(uint96);
 }
